Move `ResolveAddr` to config.go.
